# cert-api2/public/swagger.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Taichung School Certificate Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3000/api/test
  - url: http://192.168.56.101:3000/api/v1
security:
  - bearerAuth: []
paths:
...
  /dev/ethreg-contract:
    get:
      summary: "取得ORM測試合約列表"
      tags:
        - dev
      parameters: []
      responses:
        '200':
          content:
            application/json:
    post:
      summary: "新增ORM測試合約"
      tags:
        - dev
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                deployAccount:
                  type: string
              example:
                address: '0xad44a8ea9a9bb5ef66f041bb921a687331729eb4'
                deployAccount: '0x1b2ae7ea68e81998c031122ea7aa6ed0c6975ee7'
      responses:
        '200':
          content:
            application/json:
...
  /auth/oid-tcsso-authurl:
    post:
      summary: 台中市教育局校園登入 OPENID authorization_endpoint
      tags:
        - authorization
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                response_type: string
                azp: string
                redirect_uri: string
                scope: string
                state: number
              example:
                response_type: code
                scope: openid+email
      responses:
        '200':
          content:
            application/json:
...
  /auth/oid-callback:
    post:
      summary: 進行中 OpenID Connect 1.0 # 3.1.2 (http://openid.net/specs/openid-connect-core-1_0.html#AuthorizationEndpoint)
      tags:
        - authorization
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...